chore(verify_api): move token code to utils file
